Extract panel audio and styles to its own files
