+++ src/containers/nav/desktop/NavColumn.module.css
@@ -83,43 +83,6 @@ svg.logo path {
   color: var(--neutral);
 }
 
-.navColumn .audio {
-  display: flex;
-  align-items: center;
-  padding-left: 8px;
-  height: 16px;
-}
-
-.navColumn .audioAmount {
-  margin-left: 4px;
-  font-size: var(--font-s);
-  font-weight: var(--font-heavy);
-  color: var(--neutral-light-4);
-}
-
-.navColumn .earnAudio {
-  display: flex;
-  align-items: center;
-  margin-left: 12px;
-  padding: 4px 8px;
-  color: var(--white);
-  background: var(--secondary);
-  border-radius: 16px;
-  font-size: var(--font-2xs);
-  font-weight: var(--font-bold);
-  letter-spacing: .1em;
-}
-
-.navColumn .earnAudioCaret{
-  margin-left: 2px;
-  height: 12px;
-  width: 10px;
-}
-
-.navColumn .earnAudioCaret path {
-  fill: var(--white);
-}
-
 .navColumn .accountWrapper .wrapperPhoto {
   position: relative;
   cursor: pointer;
